refactor lib types
